+++ MarkdownParser/Antlr/Answer/Parts/TextPart.cs
@@ -1,4 +1,7 @@
-﻿using System.Windows.Documents;
+﻿using MarkdownParser.Helper;
+using System.Windows;
+using System.Windows.Documents;
+using System.Windows.Media.Animation;
 
 namespace MarkdownParser.Antlr.Answer.Parts
 {
@@ -47,11 +50,71 @@ namespace MarkdownParser.Antlr.Answer.Parts
 
         public IEnumerable<Inline> GetInlines(bool isInProgress)
         {
-            yield return new Run
+            const string boldAnchor = "**";
+
+            var boldIndexes = Text.AllIndexesOf(
+                boldAnchor
+                )
+                .ConvertAll(i => (Bold: true, Index: i));
+            if (boldIndexes.Count <= 1)
             {
-                FontSize = _fontSizeProvider.TextSize,
-                Text = Text
-            };
+                yield return new Run
+                {
+                    FontSize = _fontSizeProvider.TextSize,
+                    Text = Text
+                };
+                yield break;
+            }
+
+            if (boldIndexes[0].Index > 0)
+            {
+                boldIndexes.Insert(
+                    0,
+                    (Bold: false, Index: 0)
+                    );
+            }
+            if (boldIndexes.Last().Index < Text.Length - boldAnchor.Length - 1)
+            {
+                boldIndexes.Add(
+                    (Bold: false, Index: Text.Length)
+                    );
+            }
+
+            var currentFontWeight = FontWeights.Regular;
+            for (var partIndex = 0; partIndex < boldIndexes.Count - 1; partIndex++)
+            {
+                var leftPartIndex = partIndex;
+                var rightPartIndex = partIndex + 1;
+
+                var leftPair = boldIndexes[leftPartIndex];
+                var leftIndex = leftPair.Bold
+                    ? leftPair.Index + boldAnchor.Length
+                    : leftPair.Index
+                    ;
+
+                var rightPair = boldIndexes[rightPartIndex];
+                var rightIndex = rightPair.Index;
+
+                if (leftPair.Bold && rightPair.Bold)
+                {
+                    currentFontWeight =
+                        currentFontWeight == FontWeights.Bold
+                        ? FontWeights.Regular
+                        : FontWeights.Bold
+                        ;
+                }
+                else
+                {
+                    currentFontWeight = FontWeights.Regular;
+                }
+
+                yield return new Run
+                {
+                    FontSize = _fontSizeProvider.TextSize,
+                    FontWeight = currentFontWeight,
+                    Text = Text.Substring(leftIndex, rightIndex - leftIndex)
+                };
+            }
         }
     }
 }